broaden cases for checking for portal trap
Players can be considered by CB/Minecraft to be within a "semi" block,
such as glass panes or fence posts, but also still stuck in the portal
effect. Though broadening the definition could have some unintended
effects (such as if a player is near a fence when the check takes place,
but is not near the portal), it might be potentially more CPU efficient
vs. checking placement of each and every fence post, glass pane, etc.