Add restrictions to the nether portal trap check (see desc)
If the player is found to be inside a fence block (or similar) when the
portal trap check occurs, we also check this block to see if it's
adjacent to a portal block.

Since this check would only be triggered _if_ a player is inside such a
block when the portal trap detection takes place, I don't imagine this
having any significant CPU cost.